refactor(pair_match): 패키지 구조 변경
