+++ backend/src/main/java/wooteco/retrospective/domain/pair/Pairs.java
@@ -1,5 +1,8 @@
 package wooteco.retrospective.domain.pair;
 
+import wooteco.retrospective.domain.pair.member.Member;
+import wooteco.retrospective.domain.pair.member.Shuffled;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
