+++ backend/src/main/java/wooteco/retrospective/domain/pair/member/ShuffledMembers.java
@@ -1,4 +1,4 @@
-package wooteco.retrospective.domain.pair;
+package wooteco.retrospective.domain.pair.member;
 
 import java.util.ArrayList;
 import java.util.Collections;
