+++ backend/src/test/java/wooteco/retrospective/domain/pair/PairsTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import wooteco.retrospective.domain.pair.member.Member;
 
 import java.util.Collection;
 import java.util.Collections;
