+++ backend/src/test/java/wooteco/retrospective/domain/pair/ShuffledMemberTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import wooteco.retrospective.domain.pair.member.Member;
+import wooteco.retrospective.domain.pair.member.Shuffled;
+import wooteco.retrospective.domain.pair.member.ShuffledMembers;
 
 import java.util.List;
 import java.util.function.UnaryOperator;
